Update ca_env.py
Step function: 
We add 'self.old_grid' in the self.MDP function. This will help us in the render function because we need to retain the old grid in the current step. The easier way to do this is retaining in the self.state. 

We add the 'steps_beyond_done' counter. The GoL, most of the time, generates a regular pattern that never dies, so the game could never end by itself. This counter is necessary to get it done. 

Reset function: 
We add 
self.steps_beyond_done = self._max_steps
With this, we make the max_step an hyperparameter easy to change by the user